core(extension): extract a new ext-bg file, splitting extension/devtools usecases
